Continues to fumble with classes and namespaces and stuff
