+++ stream_alert/alert_processor/helpers.py
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
-
+from publishers import AlertPublisherRepository
 
 def elide_string_middle(text, max_length):
     """Replace the middle of the text with ellipses to shorten text to the desired length.
@@ -30,3 +30,77 @@ def elide_string_middle(text, max_length):
 
     half_len = (max_length - 5) / 2  # Length of text on either side.
     return '{} ... {}'.format(text[:half_len], text[-half_len:])
+
+
+def publish_alert(alert, output, descriptor):
+    """Presents the alert as a dict for output classes to send to their API integrations.
+
+    Args:
+        alert (Alert): The alert to be dispatched
+        output (OutputDispatcher): Instance of the output class dispatching this alert
+        descriptor (str): The descriptor for the output
+
+    Returns:
+        dict
+    """
+    publisher = _assemble_alert_publisher_for_output(
+        alert,
+        output,
+        descriptor
+    )
+    return publisher.publish(alert, {})
+
+
+def _assemble_alert_publisher_for_output(alert, output, descriptor):
+    """Gathers all requested publishers on the alert and returns them as a single Publisher
+
+    Note: When no publishers are requested, or when certain
+
+    Args:
+        alert (Alert): The alert that is pulled from DynamoDB
+        output (OutputDispatcher|None): Instance of OutputDispatcher that is sending the alert
+        descriptor (str): The descriptor of the Output
+
+    Returns:
+        BaseAlertPublisher
+    """
+
+    alert_publishers = alert.publishers
+    publisher_names = []
+    if isinstance(alert_publishers, basestring):
+        # Case 1: The publisher is a single string.
+        #   apply this single publisher to all outputs + descriptors
+        publisher_names.append(alert_publishers)
+    elif isinstance(alert_publishers, list):
+        # Case 2: The publisher is an array of strings.
+        #   apply all publishers to all outputs + descriptors
+        publisher_names += alert_publishers
+    elif isinstance(alert_publishers, dict):
+        # Case 3: The publisher is a dict mapping output strings -> strings or list of strings
+        #   apply only publishers under the correct output key. We look under 2 keys:
+        #   one key that applies publishers to all outputs for a specific output type, and
+        #   another key that applies publishers only to outputs of the type AND matching
+        #   descriptor.
+
+        from stream_alert.alert_processor.outputs.output_base import OutputDispatcher  # FIXME (derek.wang) figure out what to do with this inline
+
+        if isinstance(output, OutputDispatcher):
+            # Order is important here; we load the output+descriptor-specific publishers first
+            output_service_name = output.__service__
+            described_output_name = '{}:{}'.format(output_service_name, descriptor)
+            if described_output_name in alert_publishers:
+                publisher_name_or_names = alert_publishers[described_output_name]
+                if isinstance(publisher_name_or_names, list):
+                    publisher_names += publisher_name_or_names
+                else:
+                    publisher_names.append(publisher_name_or_names)
+
+            # Then load output-specific publishers second
+            if output_service_name and output_service_name in alert_publishers:
+                publisher_name_or_names = alert_publishers[output_service_name]
+                if isinstance(publisher_name_or_names, list):
+                    publisher_names = publisher_names + publisher_name_or_names
+                else:
+                    publisher_names.append(publisher_name_or_names)
+
+    return AlertPublisherRepository.create_composite_publisher(publisher_names)
