+++ stream_alert/alert_processor/outputs/phantom.py
@@ -16,7 +16,7 @@ limitations under the License.
 from collections import OrderedDict
 import os
 
-from publishers import publish_alert
+from stream_alert.alert_processor.helpers import publish_alert
 from stream_alert.alert_processor.outputs.output_base import (
     OutputDispatcher,
     OutputProperty,
