+++ stream_alert/alert_processor/outputs/slack.py
@@ -16,7 +16,7 @@ limitations under the License.
 import cgi
 from collections import OrderedDict
 
-from publishers import publish_alert
+from stream_alert.alert_processor.helpers import publish_alert
 from stream_alert.alert_processor.outputs.output_base import (
     OutputDispatcher,
     OutputProperty,
