+++ tests/unit/stream_alert_alert_processor/test_outputs/test_slack.py
@@ -18,8 +18,8 @@ from collections import Counter, OrderedDict
 from mock import patch, Mock, MagicMock
 from nose.tools import assert_equal, assert_false, assert_true, assert_set_equal
 
+from stream_alert.alert_processor.helpers import publish_alert
 from stream_alert.alert_processor.outputs.slack import SlackOutput
-from publishers import publish_alert
 from tests.unit.stream_alert_alert_processor.helpers import (
     get_random_alert,
     get_alert,
